tests: Fix error in parse_test_align filenames, path construction.

+++ tests/run.sh
@@ -380,12 +380,14 @@ function parse_test_align {
   echo -e "\t${FUNCNAME[0]}:\tparse-test-align.py ::: overlap.align.txt:"
   if ! local_prefix=$(_get_local_prefix "$cmd_prefix" tests/parse-test-align.py); then return 1; fi
   "${local_prefix}parse-test-align.py" --ref "$dirname/overlap.ref.tmp.fa" \
-    --fq1 "$dirname/overlap.sscs.tmp_1.fa" --fq2 "$dirname/overlap.sscs.tmp_2.fa" \
+    --fq1 "$dirname/overlap.reads.tmp_1.fq" --fq2 "$dirname/overlap.reads.tmp_2.fq" \
     "$dirname/overlap.align.txt" >/dev/null
-  for out in overlap.ref.tmp.fa overlap.sscs.tmp_1.fa overlap.sscs.tmp_2.fa; do
-    expected="$dirname/$(echo "$out" | sed -E 's/\.tmp//')"
+  for out in overlap.ref.tmp.fa overlap.reads.tmp_1.fq overlap.reads.tmp_2.fq; do
+    expected=$(echo "$out" | sed -E 's/\.tmp//')
     diff -s "$dirname/$expected" "$dirname/$out"
-    rm "$dirname/$out"
+    if [[ -f "$dirname/$out" ]]; then
+      rm "$dirname/$out"
+    fi
   done
 }
